gmake2: Fix detecting shell as msdos
The current method of inspecting whether the ComSpec environment variable ends
in `.exe` is not reliable, because no shell updates this, and also
`sh.exe` is a perfectly valid name for a shell running build rules.

I could not find any shell that actually changes the ComSpec environment
variable on Windows.

All of these yielded the same result

powershell: $env:ComSpec
busybox-w32: echo $COMSPEC
msys2: echo $COMSPEC

returning: C:\Windows\system32\cmd.exe

So this check doesn't work as intended, and will always return true no matter
what on Windows, even in cases where the shell the Makefile is running under is
not cmd.exe.

By default, GNU Make searches the path for an `sh.exe` and uses it if it is
present (see variable.c), before running build rules using `cmd.exe`

So, checking whether the value of SHELL ends in .exe is also not an option.

See GNU Make's README.W32 for more details on how this works.